feat: Deleting Files

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -41,6 +42,7 @@ exports.getEditProduct = (req, res, next) => {
     Product.findById(prodId)
         .then(product => {
             if (!product) {
+              fileHelper.deleteFile(product.imageUrl);
                 return res.redirect('/');
             }
             res.render('admin/edit-product', {
@@ -68,7 +70,10 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDesc;
-            product.imageUrl = updatedImageUrl;
+
+            if(image){
+              product.imageUrl = image.path;
+            }
             return product.save().then(result => {
                 console.log('UPDATED PRODUCT!');
                 res.redirect('/admin/products');
@@ -94,8 +99,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
-        .then(() => {
+    Product.findById(prodId)
+      .then(product=> {
+        if(!product){
+          return next(new Error('Product not found'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({_id: prodId, userId: req.user._id})
+      }).then(() => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
         })
